tests/wpt/web-platform-tests/fetch/metadata/sec-fetch-dest/fetch.tentative.https.sub.html

<!DOCTYPE html>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/fetch/metadata/resources/helper.js></script>
<script>
  // Site
  promise_test(t => {
    return fetch("https://{{host}}:{{ports[https][0]}}/fetch/metadata/resources/echo-as-json.py")
        .then(r => r.json())
        .then(j => assert_header_dest_equals(j, "empty"));
  }, "Same-origin fetch");

  promise_test(t => {
    return fetch("https://{{hosts[][www]}}:{{ports[https][0]}}/fetch/metadata/resources/echo-as-json.py")
        .then(r => r.json())
        .then(j => assert_header_dest_equals(j, "empty"));
  }, "Same-site fetch");

  promise_test(t => {
    return fetch("https://{{hosts[alt][www]}}:{{ports[https][0]}}/fetch/metadata/resources/echo-as-json.py")
        .then(r => r.json())
        .then(j => assert_header_dest_equals(j, "empty"));
  }, "Cross-site fetch");

  // Mode
  promise_test(t => {
    return fetch("https://{{host}}:{{ports[https][0]}}/fetch/metadata/resources/echo-as-json.py", {mode: "same-origin"})
        .then(r => r.json())
        .then(j => assert_header_dest_equals(j, "empty"));
  }, "Same-origin mode");

  promise_test(t => {
    return fetch("https://{{host}}:{{ports[https][0]}}/fetch/metadata/resources/echo-as-json.py", {mode: "cors"})
        .then(r => r.json())
        .then(j => assert_header_dest_equals(j, "empty"));
  }, "CORS mode");

  promise_test(t => {
    return fetch("https://{{host}}:{{ports[https][0]}}/fetch/metadata/resources/echo-as-json.py", {mode: "no-cors"})
        .then(r => r.json())
        .then(j => assert_header_dest_equals(j, "empty"));
  }, "no-CORS mode");
</script>
